requirements.txt
openai
python-dotenv
langchain
pypdf
langchain-text-splitters
chromadb
sentence-transformers